pyproject.toml
[project]
name = "npu-simulation"
version = "0.1.0"
description = "NPU compiler/runtime backend simulated on Metal GPU"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4",
    "ml-dtypes>=0.5.0",
    "pyobjc-framework-Metal>=12.0",
    "pyobjc-framework-MetalPerformanceShaders>=12.0",
    "pyobjc-core>=12.0",
    "pytorch-ir",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "torch>=2.10",
    "torchvision>=0.25",
    "ruff>=0.15",
    "safetensors>=0.7",
    "ty>=0.0.17",
    "pre-commit>=4.5",
]
examples = [
    "transformers>=4.50",
    "huggingface-hub>=0.30",
    "safetensors>=0.7",
    "scipy>=1.14",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.27",
    "mkdocs-static-i18n>=1.0",
]

[build-system]
requires = ["hatchling>=1.28"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["npu_compiler", "npu_runtime"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ty.environment]
python-version = "3.11"

[tool.ty.rules]
# pyobjc Metal API 타입 미지원
unresolved-attribute = "ignore"
